[osx] More alias' for docker and expose docker for mac API via http/tcp
Modified alias' for checking the docker for mac image size and removing
it.

Also, added alias for starting/exposing the docker for mac API via
http/tcp using the 'socat' command line tool. This was required for
IntelliJ integration for the docker plugin as it does not support
the UNIX socket URI scheme yet.

See http://blog.dekstroza.io/osx-docker-beta-intellj-playing-nicely/
for more details.