🐛 Add creation of target parent folder in create_symbolic_links.sh
In the script for creating symbolic links, a line was added to create the parent directory of the target file, if it does not exist already. This is to prevent errors when trying to create a symbolic link where the target folder path does not exist yet. The folder creation step is added to both 'if' and 'elif' conditions to handle all possible cases.