🔀 remove "Retina Display Menu" and add "EasyRes" in install information
The utility "Retina Display Menu" has been removed from the brew cask install list in the desktop_utilities.sh file as it is no longer available on brew. Instead, "EasyRes" has been added to the list of software to potentially install in the setup.sh file. Users will now be provided with the option to install "EasyRes" from the App Store for managing display resolutions.