👷 Add manual trigger for application installation
This commit introduces a new GitHub action for triggering manual application installations on macOS. The action takes an input parameter for selective application installation using a regex match in the name. This is to provide more flexibility in the automated development setup process. Further, commented out the previous automated install lines in macos.yml as part of transitioning into manual initiation process. This change is part of efforts to enhance control and customization over the automated build setup process.